DAS-2060: Here's a chunk of hopeful changes
