+++ xarray/backends/netCDF4_.py
@@ -8,6 +8,7 @@ from contextlib import suppress
 from typing import TYPE_CHECKING, Any
 
 import numpy as np
+from datatree import DataTree, NodePath
 
 from xarray import coding
 from xarray.backends.common import (
@@ -666,6 +667,39 @@ class NetCDF4BackendEntrypoint(BackendEntrypoint):
                 decode_timedelta=decode_timedelta,
             )
         return ds
+
+    def open_datatree(self,
+                      filename: str,
+                      **kwargs) -> DataTree:
+        from netCDF4 import Dataset as ncDataset
+
+        from xarray.backends.api import open_dataset
+
+        ds = open_dataset(filename, **kwargs)
+        tree_root = DataTree.from_dict({"/": ds})
+        with ncDataset(filename, mode="r") as ncds:
+            for path in _iter_nc_groups(ncds):
+                subgroup_ds = open_dataset(filename, group=path, **kwargs)
+
+                # TODO refactor to use __setitem__ once creation of new nodes by assigning Dataset works again
+                node_name = NodePath(path).name
+                new_node: DataTree = DataTree(name=node_name, data=subgroup_ds)
+                tree_root._set_item(
+                    path,
+                    new_node,
+                    allow_overwrite=False,
+                    new_nodes_along_path=True,
+                )
+        return tree_root
+
+
+def _iter_nc_groups(root, parent="/"):
+    parent = NodePath(parent)
+    for path, group in root.groups.items():
+        gpath = parent / path
+        yield str(gpath)
+        yield from _iter_nc_groups(group, parent=gpath)
+
 
 
 BACKEND_ENTRYPOINTS["netcdf4"] = ("netCDF4", NetCDF4BackendEntrypoint)
